requirements.txt
asn1crypto==0.24.0
cmake==3.21.3
cryptography==2.1.4
distro==1.6.0
Django==1.11.29
djangorestframework==3.9.4
enum34==1.1.6
ez-setup==0.9
idna==2.6
ipaddress==1.0.17
keyring==10.6.0
keyrings.alt==3.0
packaging==20.9
pycrypto==2.6.1
pygobject==3.26.1
pyparsing==2.4.7
pytz==2021.3
pyxdg==0.25
scikit-build==0.12.0
SecretStorage==2.3.1
six==1.11.0
